refactor: extract SVGButton component
