+++ src/index.css
@@ -98,28 +98,6 @@
   --main-song-preview-color: #051b0c;
 }
 
-/* Common Components */
-.common-svg-button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 6px;
-  font-size: 14px;
-  color: var(--spice-subtext);
-  background-color: unset;
-  border: unset;
-  cursor: pointer;
-}
-
-.common-svg-button[disabled] {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
-.common-svg-button:not([disabled]):hover {
-  color: var(--spice-text);
-}
-
 .song-preview {
   padding: 0 32px;
   display: flex;
